+++ api/src/main/java/com/illtamer/infinite/bot/api/event/request/GroupRequestEvent.java
@@ -82,7 +82,7 @@ public class GroupRequestEvent extends RequestEvent implements QuickAction {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendGroupMessage(String message) {
+    public Integer sendGroupMessage(String message) {
         return OpenAPIHandling.sendGroupMessage(message, groupId);
     }
 
@@ -90,7 +90,7 @@ public class GroupRequestEvent extends RequestEvent implements QuickAction {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendGroupMessage(Message message) {
+    public Integer sendGroupMessage(Message message) {
         return OpenAPIHandling.sendGroupMessage(message, groupId);
     }
